au3: slim import-export mod headers, make codec deps PRIVATE
Each mod-* header now declares only its plugin class (ExportXxx /
XxxImportPlugin) over ExportPlugin.h / ImportPlugin.h. The processors,
options editors, file handles, option tables and codec includes move into
the .cpp files, where their only users already were (RegisterExportPlugins /
RegisterImportPlugins only need the plugin classes).

With nothing codec-related exposed, the codec targets (FLAC, lame, mpg123,
ogg/vorbis, opus, sndfile, twolame, wavpack) and the au3 libraries that were
PUBLIC only because of those headers become PRIVATE; au3-import-export
(and au3-strings for ExportMP3.h) are the only PUBLIC deps left. CMake now
enforces that codec headers stay out of the mod interfaces.

Also adds the missing #pragma once to ImportMP3_MPG123.h, ExportWavPack.h
and ImportWavPack.h.

Part of #11803.
