+++ au3/modules/import-export/mod-flac/CMakeLists.txt
@@ -13,15 +13,15 @@ set( SOURCES
 
 set( LIBRARIES
 PUBLIC
-   FLAC::FLAC
-   FLAC::FLAC++
    au3-import-export
+PRIVATE
+   au3-math
    au3-mixer
    au3-tags
    au3-track
    au3-wave-track
-PRIVATE
-   au3-math
+   FLAC::FLAC
+   FLAC::FLAC++
    rapidjson::rapidjson
    wxBase
 )
