+++ au3/modules/import-export/mod-flac/ExportFLAC.h
@@ -4,61 +4,7 @@
 
 #pragma once
 
-#include "au3-import-export/Export.h"
-
-#include "au3-mixer/Mix.h"
-
-#include "au3-tags/Tags.h"
-#include "au3-track/Track.h"
-
-#include "au3-import-export/ExportPluginHelpers.h"
-#include "au3-import-export/PlainExportOptionsEditor.h"
-
-#include "FLAC++/encoder.h"
-
-enum : int {
-    FlacOptionIDBitDepth = 0,
-    FlacOptionIDLevel
-};
-
-struct FLAC__StreamMetadataDeleter {
-    void operator ()(FLAC__StreamMetadata* p) const
-    {
-        if (p) {
-            ::FLAC__metadata_object_delete(p);
-        }
-    }
-};
-using FLAC__StreamMetadataHandle = std::unique_ptr<
-    FLAC__StreamMetadata, FLAC__StreamMetadataDeleter
-    >;
-
-class FLACExportProcessor final : public ExportProcessor
-{
-    struct
-    {
-        TranslatableString status;
-        double t0;
-        double t1;
-        unsigned numChannels;
-        wxFileNameWrapper fName;
-        sampleFormat format;
-        FLAC::Encoder::File encoder;
-        wxFFile f;
-        std::unique_ptr<Mixer> mixer;
-    } context;
-
-public:
-
-    bool Initialize(AudacityProject& project, const Parameters& parameters, const wxFileNameWrapper& filename, double t0, double t1,
-                    bool selectedOnly, double sampleFormat, unsigned channels, MixerOptions::Downmix* mixerSpec, const Tags* tags) override;
-
-    ExportResult Process(ExportProcessorDelegate& delegate) override;
-
-private:
-
-    FLAC__StreamMetadataHandle MakeMetadata(AudacityProject* project, const Tags* tags) const;
-};
+#include "au3-import-export/ExportPlugin.h"
 
 class ExportFLAC final : public ExportPlugin
 {
