+++ au3/modules/import-export/mod-ogg/ImportOGG.cpp
@@ -38,6 +38,41 @@
  * was a bitch to track down. */
 #include <wx/ffile.h>
 
+#include "au3-import-export/Import.h"
+#include "au3-tags/Tags.h"
+#include "au3-wave-track/WaveTrack.h"
+#include "au3-import-export/ImportProgressListener.h"
+#include "au3-import-export/ImportUtils.h"
+#include <vorbis/vorbisfile.h>
+
+#define DESC TranslatableString("import-export", "Ogg Vorbis files")
+
+class OggImportFileHandle final : public ImportFileHandleEx
+{
+public:
+    OggImportFileHandle(const FilePath& filename, std::unique_ptr<wxFFile>&& file, std::unique_ptr<OggVorbis_File>&& vorbisFile);
+    ~OggImportFileHandle();
+
+    TranslatableString GetFileDescription() override;
+    double GetDuration() const override;
+    int GetRequiredTrackCount() const override;
+    ByteCount GetFileUncompressedBytes() override;
+    void Import(
+        ImportProgressListener& progressListener, WaveTrackFactory* trackFactory, TrackHolders& outTracks, Tags* tags,
+        std::optional<LibFileFormats::AcidizerTags>& outAcidTags) override;
+    wxInt32 GetStreamCount() override;
+    const TranslatableStrings& GetStreamInfo() override;
+    void SetStreamUsage(wxInt32 StreamID, bool Use) override;
+
+private:
+    std::unique_ptr<wxFFile> mFile;
+    std::unique_ptr<OggVorbis_File> mVorbisFile;
+
+    ArrayOf<int> mStreamUsage;
+    TranslatableStrings mStreamInfo;
+    std::vector<TrackListHolder> mStreams;
+};
+
 static const auto exts = {
     wxT("ogg")
 };
