+++ au3/modules/import-export/mod-pcm/CMakeLists.txt
@@ -13,17 +13,17 @@ set( SOURCES
 
 set( LIBRARIES
 PUBLIC
-   SndFile::sndfile
-   au3-file-formats
    au3-import-export
-   au3-mixer
 PRIVATE
+   au3-file-formats
    au3-files
    au3-math
+   au3-mixer
    au3-preferences
    au3-tags
    au3-track
    au3-wave-track
+   SndFile::sndfile
    wxBase
 )
