+++ au3/modules/import-export/mod-pcm/ImportPCM.h
@@ -5,7 +5,6 @@
 #pragma once
 
 #include "au3-import-export/ImportPlugin.h"
-#include "au3-file-formats/FileFormats.h"
 
 class PCMImportPlugin final : public ImportPlugin
 {
@@ -17,31 +16,4 @@ public:
     TranslatableString GetPluginFormatDescription() override;
     std::unique_ptr<ImportFileHandle> Open(
         const FilePath& Filename, AudacityProject*) override;
-};
-
-class PCMImportFileHandle final : public ImportFileHandleEx
-{
-public:
-    PCMImportFileHandle(const FilePath& name, SFFile&& file, SF_INFO info);
-    ~PCMImportFileHandle();
-
-    TranslatableString GetFileDescription() override;
-    double GetDuration() const override;
-    int GetRequiredTrackCount() const override;
-    ByteCount GetFileUncompressedBytes() override;
-    void Import(
-        ImportProgressListener& progressListener, WaveTrackFactory* trackFactory, TrackHolders& outTracks, Tags* tags,
-        std::optional<LibFileFormats::AcidizerTags>& outAcidTags) override;
-
-    wxInt32 GetStreamCount() override;
-
-    const TranslatableStrings& GetStreamInfo() override;
-
-    void SetStreamUsage(wxInt32 WXUNUSED(StreamID), bool WXUNUSED(Use)) override;
-
-private:
-    SFFile mFile;
-    const SF_INFO mInfo;
-    sampleFormat mEffectiveFormat;
-    sampleFormat mFormat;
 };
