+++ au3/modules/import-export/mod-wavpack/ImportWavPack.cpp
@@ -38,6 +38,35 @@
 
 #define DESC TranslatableString("import-export", "WavPack files")
 
+#include "au3-math/SampleFormat.h"
+
+class WavPackImportFileHandle final : public ImportFileHandleEx
+{
+public:
+    WavPackImportFileHandle(const FilePath& filename, WavpackContext* wavpackContext);
+    ~WavPackImportFileHandle();
+
+    TranslatableString GetFileDescription() override;
+    double GetDuration() const override;
+    int GetRequiredTrackCount() const override;
+    ByteCount GetFileUncompressedBytes() override;
+    void Import(ImportProgressListener& progressListener, WaveTrackFactory* trackFactory, TrackHolders& outTracks, Tags* tags,
+                std::optional<LibFileFormats::AcidizerTags>& outAcidTags) override;
+
+    wxInt32 GetStreamCount() override;
+    const TranslatableStrings& GetStreamInfo() override;
+    void SetStreamUsage(wxInt32 StreamID, bool Use) override;
+
+private:
+    WavpackContext* mWavPackContext;
+    int mNumChannels;
+    uint32_t mSampleRate;
+    int mBitsPerSample;
+    int mBytesPerSample;
+    int64_t mNumSamples;
+    sampleFormat mFormat;
+};
+
 static const auto exts = {
     wxT("wv")
 };
