+++ au3/modules/import-export/mod-wavpack/ImportWavPack.h
@@ -2,9 +2,8 @@
 * Audacity: A Digital Audio Editor
 */
 
-#include <wavpack/wavpack.h>
+#pragma once
 
-#include "au3-math/SampleFormat.h"
 #include "au3-import-export/ImportPlugin.h"
 
 class WavPackImportPlugin final : public ImportPlugin
@@ -17,31 +16,4 @@ public:
     TranslatableString GetPluginFormatDescription() override;
     std::unique_ptr<ImportFileHandle> Open(
         const FilePath& Filename, AudacityProject*) override;
-};
-
-class WavPackImportFileHandle final : public ImportFileHandleEx
-{
-public:
-    WavPackImportFileHandle(const FilePath& filename, WavpackContext* wavpackContext);
-    ~WavPackImportFileHandle();
-
-    TranslatableString GetFileDescription() override;
-    double GetDuration() const override;
-    int GetRequiredTrackCount() const override;
-    ByteCount GetFileUncompressedBytes() override;
-    void Import(ImportProgressListener& progressListener, WaveTrackFactory* trackFactory, TrackHolders& outTracks, Tags* tags,
-                std::optional<LibFileFormats::AcidizerTags>& outAcidTags) override;
-
-    wxInt32 GetStreamCount() override;
-    const TranslatableStrings& GetStreamInfo() override;
-    void SetStreamUsage(wxInt32 StreamID, bool Use) override;
-
-private:
-    WavpackContext* mWavPackContext;
-    int mNumChannels;
-    uint32_t mSampleRate;
-    int mBitsPerSample;
-    int mBytesPerSample;
-    int64_t mNumSamples;
-    sampleFormat mFormat;
 };
